# docs/swagger.yaml
definitions:
...
  models.UserSkillProficiency:
    enum:
    - beginner
    - intermediate
    - advanced
    - expert
    type: string
    x-enum-varnames:
    - ProficiencyBeginner
    - ProficiencyIntermediate
    - ProficiencyAdvanced
    - ProficiencyExpert
...
  ports.CreateUserSkillInput:
    properties:
      proficiency_level:
        allOf:
        - $ref: '#/definitions/models.UserSkillProficiency'
        enum:
        - beginner
        - intermediate
        - advanced
        - expert
      skill_id:
        type: integer
      user_id:
        type: integer
    required:
    - proficiency_level
    - skill_id
    - user_id
    type: object
...
  ports.SkillResponse:
    properties:
      active:
        type: boolean
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  ports.SubscriptionResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      valid_days:
        type: integer
      valid_months:
        type: integer
    type: object
...
  ports.UpdateUserSkillInput:
    properties:
      proficiency_level:
        allOf:
        - $ref: '#/definitions/models.UserSkillProficiency'
        enum:
        - beginner
        - intermediate
        - advanced
        - expert
    type: object
...
  ports.UserResponse:
    properties:
      active:
        type: boolean
      contact_email:
        type: string
      contact_phone_no:
        type: string
      created_at:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      login_email:
        type: string
      updated_at:
        type: string
    type: object
  ports.UserSkillResponse:
    properties:
      created_at:
        type: string
      proficiency_level:
        $ref: '#/definitions/models.UserSkillProficiency'
      skill:
        $ref: '#/definitions/ports.SkillResponse'
      skill_id:
        type: integer
      user:
        $ref: '#/definitions/ports.UserResponse'
      user_id:
        type: integer
    type: object
  ports.UserSkillsResponse:
    properties:
      count:
        type: integer
      skills:
        items:
          $ref: '#/definitions/ports.UserSkillResponse'
        type: array
    type: object
...
  ports.UserSubscriptionResponse:
    properties:
      date_from:
        type: string
      date_to:
        type: string
      id:
        type: integer
      is_trial:
        type: boolean
      subscription:
        $ref: '#/definitions/ports.SubscriptionResponse'
      user:
        $ref: '#/definitions/ports.UserResponse'
    type: object
...
info:
  contact: {}
paths:
...
  /users/{id}/skills:
    get:
      description: Retrieves all skills and proficiency levels associated with the
        user. Enforces self-management.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of user skills
          schema:
            $ref: '#/definitions/ports.UserSkillsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Not the target user)
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get User Skills
      tags:
      - users
      - skills
    post:
      consumes:
      - application/json
      description: Adds a new skill and its proficiency level to the authenticated
        user's profile. Enforces self-management.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Skill details (SkillID, ProficiencyLevel)
        in: body
        name: skill
        required: true
        schema:
          $ref: '#/definitions/ports.CreateUserSkillInput'
      produces:
      - application/json
      responses:
        "201":
          description: Skill added successfully
          schema:
            $ref: '#/definitions/ports.UserSkillResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Not the target user)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ErrUserNotFound or ErrSkillNotFound
          schema:
            additionalProperties: true
            type: object
        "409":
          description: ErrUserSkillAlreadyExists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add User Skill
      tags:
      - users
      - skills
  /users/{id}/skills/{skillID}:
    delete:
      description: Removes a skill association from the user's profile. Enforces self-management.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Skill ID
        in: path
        name: skillID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Skill removed successfully (No Content)
        "400":
          description: Invalid ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Not the target user)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ErrUserSkillNotFound
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove User Skill
      tags:
      - users
      - skills
    put:
      consumes:
      - application/json
      description: Updates the proficiency level for an existing skill associated
        with the user. Enforces self-management.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Skill ID
        in: path
        name: skillID
        required: true
        type: integer
      - description: New proficiency level
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/ports.UpdateUserSkillInput'
      produces:
      - application/json
      responses:
        "200":
          description: Proficiency updated successfully
          schema:
            $ref: '#/definitions/ports.UserSkillResponse'
        "400":
          description: Invalid ID, request body, or validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Not the target user)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ErrUserSkillNotFound
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update User Skill Proficiency
      tags:
      - users
      - skills
  /users/{id}/subscriptions:
    get:
      description: Retrieves all currently active user subscription records (those
        not yet expired). Enforces self-management.
      parameters:
      - description: Target User ID (must match authenticated user)
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of active user subscriptions
          schema:
            items:
              $ref: '#/definitions/ports.UserSubscriptionResponse'
            type: array
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: Cannot view another user''s subscriptions'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Active Subscriptions for User
      tags:
      - users
      - subscriptions
  /users/{id}/subscriptions/{userSubscriptionID}:
    delete:
      description: Cancels a specific user subscription record by deleting it from
        the database. Enforces self-management and ownership of the record.
      parameters:
      - description: Target User ID (must match authenticated user)
        in: path
        name: id
        required: true
        type: integer
      - description: User Subscription Record ID to cancel
        in: path
        name: userSubscriptionID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Subscription cancelled successfully (No Content)
        "400":
          description: Invalid user or subscription ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 'Forbidden: You are not the owner of this record'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: ErrUserSubscriptionNotFound
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal error during cancellation
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cancel User Subscription
      tags:
      - users
      - subscriptions
swagger: "2.0"